backend/requirements.txt
fastapi
uvicorn[standard]
google-genai
anthropic
python-dotenv
pydantic
httpx
playwright
aiofiles
